+++ src/server/shared/Packets/WorldPacket.h
@@ -31,24 +31,24 @@ class WorldPacket : public ByteBuffer
         WorldPacket()                                       : ByteBuffer(0), m_opcode(0)
         {
         }
-        explicit WorldPacket(uint16 opcode, size_t res=200) : ByteBuffer(res), m_opcode(opcode) { }
+        explicit WorldPacket(uint32 opcode, size_t res=200) : ByteBuffer(res), m_opcode(opcode) { }
                                                             // copy constructor
         WorldPacket(const WorldPacket &packet)              : ByteBuffer(packet), m_opcode(packet.m_opcode)
         {
         }
 
-        void Initialize(uint16 opcode, size_t newres=200)
+        void Initialize(uint32 opcode, size_t newres=200)
         {
             clear();
             _storage.reserve(newres);
             m_opcode = opcode;
         }
 
-        uint16 GetOpcode() const { return m_opcode; }
-        void SetOpcode(uint16 opcode) { m_opcode = opcode; }
+        uint32 GetOpcode() const { return m_opcode; }
+        void SetOpcode(uint32 opcode) { m_opcode = opcode; }
 
     protected:
-        uint16 m_opcode;
+        uint32 m_opcode;
 };
 #endif
